css 040

+++ js/apps/mister-email/cmps/email-short-details.cmp.js
@@ -10,9 +10,9 @@ export default {
     template:`
     <section class="emails-short-details-container">
         <div class="btn-short-details">
-            <router-link v-if="email.type!== 'draft'" :to="'email/'+email.id"><button v-if="checkEmailType">✉</button></router-link>
-            <button v-if="email.type!== 'draft'" @click="emailToKeep(email)">✒</button>
-            <button @click="deleteEmail(email.id)">🗑</button>
+            <router-link v-if="email.type!== 'draft'" :to="'email/'+email.id"><button v-if="checkEmailType" class="email-read-details">⬜</button></router-link>
+            <button v-if="email.type!== 'draft'" @click="emailToKeep(email)" class="email-note-details">📝</button>
+            <button @click="deleteEmail(email.id)" class="email-delete-details">🗑</button>
         </div>
         <p class="short-subject">{{email.subject}}</p>
         <!-- <p>{{email.body}}</p> -->
@@ -21,11 +21,31 @@ export default {
     ,
     methods:{
         deleteEmail(emailId){
-            emailService.deleteMail(emailId)
-                .then(emails=>{
+            var prmUserDecision = Swal.fire({
+                title: 'Delete this email?',
+                icon: 'warning',
+                showCancelButton: true,
+              })
+              
+              prmUserDecision.then(res => {
+                if (res.value) {
+                  Swal.fire(
+                    'Deleted!',
+                    'Your Email has been deleted.',
+                    'success'
+                  )
+                  emailService.deleteMail(emailId)
+                .then(()=>{
                     this.$emit('deleted',true)
                     eventBus.$emit('update-percent');
                 });
+                }
+              })
+            // emailService.deleteMail(emailId)
+            //     .then(emails=>{
+            //         this.$emit('deleted',true)
+            //         eventBus.$emit('update-percent');
+            //     });
         },
         emailToKeep(email){
             emailService.saveEmailToStorage(email)
